requirements.txt
flask
langchain
langchain-community
langchain-huggingface
transformers
sentence-transformers
faiss-cpu
docx2txt
python-dotenv
pinecone-client
langchain-pinecone
